feat: Move fee-related entities and update imports
