+++ src/datasources/fee-service-api/fee-service-api.service.ts
@@ -14,12 +14,12 @@ import {
 import { CacheRouter } from '@/datasources/cache/cache.router';
 import { ILoggingService, LoggingService } from '@/logging/logging.interface';
 import { RelayFeeConfiguration } from '@/modules/relay/domain/entities/relay.configuration';
-import type { TxFeesRequest } from '@/modules/transactions/domain/entities/relay-fee/tx-fees-request.dto';
-import type { TxFeesResponse } from '@/modules/transactions/domain/entities/relay-fee/tx-fees-response.dto';
+import type { TxFeesRequest } from '@/modules/fees/domain/entities/tx-fees-request.dto';
+import type { TxFeesResponse } from '@/modules/fees/domain/entities/tx-fees-response.dto';
 import {
   CanRelayResponseSchema,
   TxFeesResponseSchema,
-} from '@/modules/transactions/domain/entities/relay-fee/schemas/tx-fees-response.schema';
+} from '@/modules/fees/domain/entities/schemas/tx-fees-response.schema';
 import type { Address } from 'viem';
 
 @Injectable()
