+++ src/modules/fees/domain/entities/tx-fees-request.dto.ts
