+++ src/modules/fees/domain/entities/tx-fees-response.dto.ts
